+++ backend/src/main/java/com/ssafy/forest/repository/ArticleCommentReplyRepository.java
@@ -13,6 +13,6 @@ public interface ArticleCommentReplyRepository extends JpaRepository<ArticleComm
     Page<ArticleCommentReply> findAllByArticleCommentOrderByCreatedAt(Pageable pageable,
         ArticleComment articleComment);
 
-    int countByArticleCommentId(long commentId);
+    long countByArticleCommentId(long commentId);
 
 }
